取消挂机按钮逻辑适配最新thread方式

+++ src/main/python/main.py
@@ -1,5 +1,5 @@
 from fbs_runtime.application_context.PyQt5 import ApplicationContext
-from PyQt5.QtCore import QRunnable,QThread,pyqtSlot
+from PyQt5.QtCore import QThread
 import sys
 from PyQt5.QtWidgets import QMainWindow, QMessageBox, QDialog
 from ui_yyswindow import Ui_MainWindow
@@ -10,9 +10,6 @@ from game import soul
 
 from configparser import ConfigParser
 import importlib
-import os
-import signal
-import time
 
 class AppContext(ApplicationContext):
 
@@ -280,45 +277,32 @@ class YYSWindow(QMainWindow):
             importlib.reload(game_func)
 
             # 防止卡死，将soul进程单独fork出子进程
-            # p = multiprocessing.Process(target=soul,kwargs={'dry_run':False})
             self.p = MyThread(fn=soul, focus=focus, exec_count=exec_count, team_leader=team_leader, auto=auto, reward=reward, dry_run=False)
-            # p = Worker(fn=soul,kwargs={'focus':focus, 'exec_count':exec_count, 'team_leader':team_leader, 'auto':auto, 'reward':reward, 'dry_run':False})
-            # p = multiprocessing.Process(target=soul,kwargs={'focus':focus, 'exec_count':exec_count, 'team_leader':team_leader, 'auto':auto, 'reward':reward, 'dry_run':False})
-            # p.start
-            # self.p = MyThread(parent=self)
             self.p.start()
 
-
-            # 将子进程赋予self.child_pid变量
-            # self.child_pid = p.pid
             self.__ui.btn_soul_start.setText('挂机中...')
             self.lock_items_soul()
             self.__ui.btn_soul_start.setEnabled(False)
             self.__ui.btn_soul_stop.setText('停止挂机')
 
     def stop_soul(self):
-        self.p.terminate()
-        self.p.wait()
-        print(1111)
-        if self.__ui.btn_soul_stop.text() == '停止挂机':
-            cfg.set('execute', 'exec_flag', '0')
-            with open(config_file, 'w', encoding='utf-8') as configfile:
-                cfg.write(configfile)
 
-            if self.child_pid is None:
+        if self.__ui.btn_soul_stop.text() == '停止挂机':
+            try:
+                if self.p.isFinished():
+                    self.showMessageBox(title='提示', message='当前没有执行挂机任务', icon=QMessageBox.Information)
+                else:
+                    self.p.terminate()
+                    self.p.wait()
+                    cfg.set('dry_run', 'flag', '1')
+                    with open(config_file, 'w', encoding='utf-8') as configfile:
+                        cfg.write(configfile)
+                    self.__ui.btn_soul_start.setText('配置检测')
+                    self.release_items_soul_stop()
+                    self.showMessageBox(title='提示', message='已停止挂机', icon=QMessageBox.Information)
+            except AttributeError:
                 self.showMessageBox(title='提示', message='当前没有执行挂机任务', icon=QMessageBox.Information)
-            else:
-                try:
-                    os.kill(self.child_pid,signal.SIGTERM)
-                except OSError:
-                    pass
-                self.child_pid = None
-                cfg.set('dry_run','flag','1')
-                with open(config_file, 'w', encoding='utf-8') as configfile:
-                    cfg.write(configfile)
-                self.showMessageBox(title='提示', message='已停止挂机', icon=QMessageBox.Information)
-                self.__ui.btn_soul_start.setText('配置检测')
-                self.release_items_soul_stop()
+
         elif self.__ui.btn_soul_stop.text() == '解除配置锁定':
             self.release_items_soul()
             self.__ui.btn_soul_start.setText('配置检测')
@@ -351,17 +335,17 @@ class YYS_pos_config(QDialog):
         self.__ui.btn_ok.clicked.connect(self.close)
 
 class MyThread(QThread):
+    '''
+    通用型Thread入口
+    '''
     def __init__(self,fn, *args, **kwargs):
-    # def __init__(self, *args, **kwargs):
         QThread.__init__(self)
         self.fn = fn
         self.args = args
         self.kwargs = kwargs
 
-    # @pyqtSlot()
     def run(self):
-        # print(111)
-        # time.sleep(5)
+
         self.fn(*self.args, **self.kwargs)
 
 
@@ -376,8 +360,6 @@ if  __name__ == "__main__":
     config_file = yys_config_path
     cfg.read(config_file, encoding='utf-8')
 
-    # my_thread = QThread()
-    # my_thread.start()
     exit_code = appctxt.run()
     sys.exit(exit_code)
